Fix url route and module path to the JS script for current project
This project doesn't really need any JS, but it would be cool if there
was a validation for the password and confirm password controls.

I've also added some comments in the routes.js file explaining what
route and module stand for (in case I become indisposed again in case of
another accident (though I'd rather not...) and forget how my own code
works) and why there's a separate variable pointing to the directory with
Intermediate HTML CSS course JS files.