Add header section to the Admin Dashboard project
The header is filled with content according to The Odin Project's
design image (but obviously it will be styled differently).

Styling of the section is not yet finished.

Added main content section to the project, which is currently empty,
but it has some placeholder background color so as to see how much
space does it occupy.