Finish (most likely) Knights Travails project.
It seems to be working. Went the simplistic route for this one: the task
was to output an array of coordinates the chess Knight has to step on
to move from the start to the finish and it does that and just that.
The gameboard wasn't needed to calculate the shortest route and thus it's
not implemented. It would be needed if e.g. obstacles were present.