Use static variables for tasks' and projects' id
Now the ID of the next Project and the next Task is stored as a static
variable in the Project and Tasks classes respectively. On creation of
an instance of either of the classes it sets the instance's ID to equal
to the variable and THEN it increments the variable by 1.