Experiment with styling the Project Form
Most of the time was wasted trying to understand why using @use instead
of @import in the scss files does not work, at least with sass-loader,
despite @import being deprecated in favor of @use.

There are also minor changes to the Project Form's structure, including
default props. I'm not sure if I want to reuse the form for editing
existing projects (which might require refactoring) or write another
one which will fit the tabular structure of the projects' list (it's
implemented as a table for now) or to refactor the projects' list not to
be a <table>, but a regular div soup (which seems to be a bad idea, since
a table seems to be appropriate for such a kind of data).