Add EditsContext
This has one purpose: to keep the counter of how many edit forms are opened
at the moment (both for project and task tables). If there's even only
one form opened (a row is set to edit mode), the edits counter can be
used to stop the user from accidentally visiting a task or project before
closing all the forms (via either discarding or saving the changes).