A few thing added
Bubbles are now fully functional (both styling and interactivity work)

Some parts of code have been refactored to make them less of a pain
to work with (this task is not yet fully finished, but since it's a pet
project it's not THAT important)

Added a separate carousel with 4 images (which show a bit more than
a single colour). The goal is to demonstrate that the JS code is reusable,
as long as proper HTML structure is maintained.